requirements.txt
pyautogui
python-xlib
Pillow
tenacity
python-dotenv
langgraph
langchain_openai
langchain-core
fastapi
uvicorn[standard]
websockets
openai